Refactor tehcnologies and languages code
