+++ src/utils/language-to-logo.ts
@@ -6,7 +6,7 @@ export const languageToLogo: { [key: string]: EThirdPartyLogoVariant } = {
     'C++': EThirdPartyLogoVariant.C_PLUS_PLUS,
     'C#': EThirdPartyLogoVariant.C_SHARP,
     Go: EThirdPartyLogoVariant.GO,
-    Java: EThirdPartyLogoVariant.JAVA,
+    Java: EThirdPartyLogoVariant.JAVA_LOGOMARK,
     JavaScript: EThirdPartyLogoVariant.JAVASCRIPT,
     PHP: EThirdPartyLogoVariant.PHP,
     Python: EThirdPartyLogoVariant.PYTHON,
